requirements.txt
tensorflow==2.6.0
matplotlib==3.4.2
streamlit==0.84.2
numpy==1.20.3
opencv_python==4.5.3.56
Pillow==8.3.2
